requirements.txt
Flask==2.3.3
flask-cors==4.0.0
websocket-client==1.6.3
python-dotenv==1.0.0
